+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 3.0.0-dev
 - possible BC break: conditions now throw TypeError instead of InvalidArgumentException
 - BC break: added type hint for parameter $parameter of method BookPage::addCondition
 - BC break: removed support for Nette style events
+- BC break: BookControl::$pages now accepts Closure or BookPagesStorage
 
 Version 2.5.1
 - allowed nexendrie/utils 2